requirements.txt
pandas>=1.3.0
numpy>=1.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=1.10.0
scikit-learn>=1.0.0
openai>=1.0.0
loguru>=0.6.0
tqdm>=4.62.0
elasticsearch>=7.0.0
datasets>=2.0.0
transformers>=4.20.0
jsonlines>=3.0.0
sqlalchemy>=1.4.0
sqlitedict>=2.0.0